Behaviour of file name casing returned by automation model seems to have changed
There were certain conditions where <see cref="Document"/> provided <see cref="Document.FullName"/> in lower case. When a project item was renamed, its full name was provided with the correct casing. In order for the names to match, it needed to be compared against <see cref="DocumentMetadata.CorrectedFullName"/>

The casing for each file now appears to be consistent